Separate sound playback into backends; implement sox backend
Currently this selects the alphabetically first backend to report itself
as available. This should be changed later in favor of user preferences.
